firstChar is no longer faster
I had to grab a bigger diff to even get a result greater than 1ms. The
performance is too close to call either way, but this is more readable I
think.

Note it's not semantically equivalent. I made some assumptions that if
they're not true may cause trouble. Should find (or make up) a diff that
exercises all the angles of the code as well as testing performance.